Debug Terrain Generation, Added Neighbor Method, Added Debug scripts
Tried to Fix several Errors with Terrain Generation:
Terrain came out crooked(because of modulo edgecase); Neighborscript didn't work correctly, went over coordinate-math and fixed it.
Added finished neighbor method to survey tile-neighbors.
Because of Debugging, created two(2) Gamescripts to write Tile coordinates to a Canvas UI and place it over Tile for easier in-game debugging.(currently not attached to any Tileblocks, but semi-easy to add)